==> 02-data.Rmd <==
# Data sources

Four datasets are collected from NYC OpenData, a free public data published by New York City agencies and other organizations. We chose this website because it provides the public with reliable data sources used by New York City government. The datasets include Free Residential at-the-tap Lead and Copper Data, Compliance at-the-tap Lead and Copper Data,Children Tested for Lead by Age 3, and Children Under 6 yrs with Elevated Blood Lead Levels (BLL). The two datasets on lead and copper data are provided by Department of Environmental Protection (DEP), and the other two on children are provided by Department of Health and Mental Hygiene. These datasets help examine the tap water quality in NYC and how children are affected by lead in NYC.

## Free Residential at-the-tap Lead and Copper Data 
It is created on February 23, 2018 and last updated on February 8, 2020. Data are collected through Free Residential (FR) at-the-tap lead testing program, which is offered to any resident in NYC upon request. Residents collect all samples using a sampling kit provided by DEP and each record provides the at-the-tap lead and copper concentrations for a specific location collected on the data noted. The variables of the dataset include kit ID, borough and zipcode where sample was collected, date collected, date received by DEP, concentration of lead in the first draw sample, in the 1-2 minute flush sample, and in the 5 minute flush sample, concentration of copper in the first draw sample, in the 1-2 minute flush sample, and in the 5 minute flush sample. Concentration is reported as milligrams per liter (mg/L). There are 19k records in this dataset. Some potential issues with the data are: 1. The concentration of lead and copper in 5 minute flush sample has lots of missing values, which impedes accurate analysis; 2. The locations of samples collected are not precise enough for further analysis.

## Compliance at-the-tap Lead and Copper Data
It is created on February 24, 2018 and last updated on February 8, 2020. Data are collected under the federal Lead and Copper Rule for compliance. Residents collect all samples using a sampling kit provided by DEP and each record provides the at-the-tap lead and copper concentrations for a specific location collected on the data noted to assess the effectiveness of corrosion control. The variables include kit ID, borough and zipcode of location where sample was collected, date collected, date received by DEP, concentration of lead in the first draw sample, and concentration of copper in the first draw sample. Concentration is reported as micrograms per liter (\mu g/L). There are 2460 records in the dataset. There is one wrong record of borough in the dataset, which needs to be ignored during analysis.

## Children Tested for Lead by Age 3 
It is created on July 10, 2018 and last updated on February 8, 2020. These data are an indicator of the number and percentage of children turning 3 years old in a given year who were tested for lead poisoning. The variables are geo_type, geo_area_id, geo_area_name, borough_id, time_period, children tested for lead by age 3 years Number, corresponding notes, children tested for lead by age 3 years Percentage, and corresponding notes. The geo_type is the type of geographic aggregation for the data, including neighborhood, borough, and citywide. The geo_area_id is an ID for the geographic neighborhood, and geo_area_name is the neighborhood name for the geographic division. Among the variables, geo_area_id, children tested for lead by age 3 years Number and Percentage are numbers, while the rest are all plain text. There are 432 records in this dataset. 

## Children Under 6 yrs with Elevated Blood Lead Levels (BLL)
It is created on July 10, 2018 and last updated on February 8, 2020. These data are an indicator of children younger thant 6 years old tested in NYC in a given year with blood lead levels of 5 mcg/dL or greater. The variables are geo_type, geo_area_id, geo_area_name, borough_id, time_period, children under 6 years with elevated blood lead levels (BLL) Number BLL >= 5\mu g/dL, Number BLL >= 10\mu g/dL, Number BLL >= 15\mu g/dL, Number Tested, Rate BLL >= 5\mu g/dL per 1,000 tested, Rate BLL >= 10\mu g/dL per 1,000 tested, Rate BLL >= 15\mu g/dL per 1,000 tested, and corresponding notes. The geo_type, geo_area_id, geo_area_name, borough_id, time_period are same as those in `Children Tested for Lead by Age 3`. There are 576 records in the dataset. One potential issue is the bias caused by total number of children tested vary a lot across NYC.
