==> 03-cleaning.Rmd <==
# Data transformation

```{r}
library(sqldf)
```

## Data loading and cleaning

Datasets are first downloaded as csv files and then read into R. 

```{r}
free_residential <- read.csv("Free_Residential_at-the-tap_Lead_and_Copper_Data.csv")
compliance <- read.csv("Compliance_at-the-tap_Lead_and_Copper_Data.csv")
children_6_yrs <- read.csv("Children_Under_6_yrs_with_Elevated_Blood_Lead_Levels__BLL_.csv")
children_by_3 <- read.csv("Children_Tested_for_Lead_by_Age_3.csv")
```

In the two datasets for at-the-tap water test results, the "Borough" column contain both uppercase and lowercase strings while refering to the same borough, hence we transformed all rows of "Borough" column to be uppercase strings.

```{r}
free_residential$Borough <- toupper(free_residential$Borough)
compliance$Borough <- toupper(compliance$Borough)
```

The units of at-the-tap lead level are different in two datasets. Hence, we transform the unit in "compliance" data from $\mu g/l$ to $mg/l$.

```{r}
compliance$First.Draw.at.the.tap.Lead.level..µg.l. <- 0.001 * compliance$First.Draw.at.the.tap.Lead.level..µg.l.
colnames(compliance)[6] <- "First.Draw.at.the.tap.Lead.level..mg.l."
```

## Data preprocessing

Since two at-the-tap lead level datasets have many variables in common, we pick out their common columns and merge them to be a single large table. Meanwhile, we add an extra column to denote which dataset the observation is originally from.

```{r}
len_fr <- length(free_residential$Kit.ID)
len_cp <- length(compliance$KIT.ID.Number)
from_fr <- c()
over_threshold_fr <- c()
for (i in 1:len_fr){
  from_fr<-c(from_fr,"free residential")
  over_threshold_fr<-c(over_threshold_fr, free_residential[i, colnames(free_residential)[6]] > 0.015)
}
free_residential$source <- from_fr
free_residential$over_thresold <- over_threshold_fr
from_cp <- c()
over_threshold_cp <- c()
for (i in 1:len_cp){
  from_cp<-c(from_cp,"compliance")
  over_threshold_cp<-c(over_threshold_cp, compliance[i, colnames(compliance)[6]] > 0.015)
}
compliance$source <- from_cp
compliance$over_threshold <- over_threshold_cp
colnames(free_residential)[c(1,2,3,4,5,6,9,12,13)] <- colnames(compliance)
merged_col <- colnames(free_residential)[c(1,2,3,4,5,6,9,12,13)]
at_the_tap_merged <- rbind(compliance, free_residential[merged_col])
write.csv(free_residential, "free_residential.csv")
write.csv(compliance, "compliance.csv")
```

## Compute ratio of samples exceeding safety threshold and group by zipcode

According to United States Environmental Protection Agency (EPA), a lead level over 0.015 mg/L is unsafe for human body. Hence, we add a binary label to show the safety status of the water sample and calculate the rate of having unsafe samples in each area.

```{r}
rate_by_zipcode <- sqldf('SELECT 
                         Zipcode, source, AVG(over_threshold) as rate 
                         FROM at_the_tap_merged 
                         GROUP BY Zipcode, source')
write.csv(rate_by_zipcode, 'tap_rate_by_zipcode.csv')
avg_by_zipcode <- sqldf('SELECT 
                         Zipcode, source, AVG(`First.Draw.at.the.tap.Lead.level..mg.l.`) as avg_lead_level
                         FROM at_the_tap_merged 
                         GROUP BY Zipcode, source')
write.csv(avg_by_zipcode, 'tap_avg_by_zipcode.csv')
```

## Join at-the-tap water data and children data

Since the water dataset contains zipcode information and the children data contains name of the neighbourhoods, we make a dictionary by looking up the zipcodes on official websites and join the two datasets using the dictionary. In this way, we are able to explore the relationship between the rate of having unsafe water samples and the rate of having unhealthy children blood samples.

```{r}
dict <- read.csv("area_names.csv")
children_rate_by_area_year <- sqldf("SELECT 
                                    time_period, geo_area_name, borough_id,  `Children.under.6.years.with.elevated.blood.lead.levels..BLL..Rate..BLL..5.µg.dL.per.1.000.tested` AS children_rate
                                    FROM children_6_yrs 
                                    WHERE geo_type = 'Neighborhood (UHF 42)' 
                                    AND time_period > 2013
                                    ORDER BY geo_area_name")
water_rate_by_area <- sqldf("SELECT 
                            Name, AVG(over_threshold) as rate                       
                            FROM (
                              SELECT 
                                compliance.*, dict.Name
                              FROM compliance
                              LEFT JOIN dict
                              ON compliance.Zipcode = dict.Zipcode
                              WHERE dict.Name is not NULL
                              AND CAST(SUBSTR(compliance.`Date.Collected`, 7, 4) AS int) <= 2016) x
                            GROUP BY Name
                            ")
join_all <- sqldf("SELECT a.time_period, a.geo_area_name, a.borough_id, a.children_rate, b.rate 
                  FROM children_rate_by_area_year a 
                  LEFT JOIN water_rate_by_area b 
                  ON a.geo_area_name = b.Name
                  WHERE b.Name is not null")
scatter <- sqldf("SELECT 
                 a.geo_area_name, a.borough_id, a.rate as rate_14, a.children_rate as crate_14, b.rate as rate_15, b.children_rate as crate_15, c.rate as rate_16, c.children_rate as crate_16
                 FROM (SELECT * FROM join_all WHERE time_period = 2014) a 
                 LEFT JOIN (SELECT * FROM join_all WHERE time_period = 2015) b 
                 ON a.geo_area_name = b.geo_area_name 
                 LEFT JOIN (SELECT * FROM join_all WHERE time_period = 2016) c 
                 ON a.geo_area_name = c.geo_area_name 
                 ")
write.csv(scatter, "scatter_data.csv")
```
