==> 05-results.Rmd <==
# Results
```{r}
library(tidyverse)
library(dplyr)
library(ggplot2)
library(ggridges)
```

## How prevalent is lead in NYC water system?
```{r, warning=FALSE, message=FALSE}
free_residential <- read.csv("free_residential.csv")
compliance <- read.csv("compliance.csv")
children_6_yrs <- read.csv("Children_Under_6_yrs_with_Elevated_Blood_Lead_Levels__BLL_.csv")
lead_free <- free_residential %>%
  select(Date.Collected, First.Draw.at.the.tap.Lead.level..mg.l., Lead.1.2.Minute.Flush..mg.L.) %>%
  mutate(year = substr(Date.Collected, 7, 10), over_threshold = (First.Draw.at.the.tap.Lead.level..mg.l. > 0.015 |  Lead.1.2.Minute.Flush..mg.L. > 0.015)) %>%
  select(year, over_threshold) %>%
  group_by(year, over_threshold) %>%
  summarise(count = n()) %>%
  na.omit %>%
  group_by(year) %>%
  mutate(total = sum(count)) %>%
  filter(over_threshold == TRUE) %>%
  mutate(ratio = count / total) %>%
  select(year, ratio) %>%
  mutate(type = "Percentage of water samples containing lead over threshold(Free-residential)")

lead_compliance <- compliance %>%
  select(Date.Collected, First.Draw.at.the.tap.Lead.level..mg.l.) %>%
  mutate(year = substr(Date.Collected, 7, 10), over_threshold = (First.Draw.at.the.tap.Lead.level..mg.l. > 0.015 )) %>%
  select(year, over_threshold) %>%
  group_by(year, over_threshold) %>%
  summarise(count = n()) %>%
  na.omit %>%
  group_by(year) %>%
  mutate(total = sum(count)) %>%
  filter(over_threshold == TRUE) %>%
  mutate(ratio = count / total) %>%
  select(year, ratio) %>%
  mutate(type = "Percentage of water samples containing lead over threshold(Compliance)")

combined <- rbind(lead_free, lead_compliance)
ggplot(data = combined, aes(fill= type, x = year, y=ratio)) +
 geom_bar(stat="identity", position = "dodge") +
 theme(legend.position="bottom") +
 guides(fill = guide_legend(nrow = 2)) +
 ggtitle("Prevalence of Lead in At-the-tap Water") +
 theme(plot.title = element_text(lineheight=.8, face="bold"))
```

We studied two datasets for prevalence of lead in at-the-tap water. One is from water samples tested under compliance, marked as red, and the other is tested upon residents' request, marked as blue.

According to EPA(Environmental Protection Agency), the maximum allowable concentration of lead in drinking water is 15 µg/L. The grouped bar chart shows the percentage of water sample containing lead over the threshold(15 µg/L) every year from 2014 to 2019. From the chart, we can observe that percentage of water samples containing lead over15 µg/L from both compliance and residents decreased by year. The decrease of contaminated percentage is more remarkable in free-residential water samples, with almost half drop from year 2015 to year 2016. The reason might be that the government has been addressing the issue of lead-contaminated water more seriously and the technology improved the effect of drinking water purifying. However, by the end of year 2019, the percentage of water samples containing lead over the threshold from residents is still over 3.8% and that from compliance is even higher, over 5%. The issue of lead-contaminated water still presents and it poses a strong threat to citizens' health.


```{r, warning=FALSE, message=FALSE, results=FALSE}
lead_con <- free_residential %>%
  select( First.Draw.at.the.tap.Lead.level..mg.l., Lead.1.2.Minute.Flush..mg.L.) %>%
  na.omit %>%
  rename(first.draw = First.Draw.at.the.tap.Lead.level..mg.l., "1-2.minute.draw" = Lead.1.2.Minute.Flush..mg.L. ) %>%
  pivot_longer(cols = everything(), names_to = "Density", values_to = "Lead Concentration(mg/L)") %>%
  filter(`Lead Concentration(mg/L)` < 0.01)

ggplot(lead_con, aes(x = `Lead Concentration(mg/L)`, y = Density)) + 
  geom_density_ridges_gradient(fill = "#00AFBB") +
  theme_ridges() +
  ggtitle("Compare Lead Concentration between First-draw Sample \n and 1-2-minute-flush Sample") +
  theme(plot.title = element_text(lineheight=.8, face="bold", hjust = 0.5))
  

```

Since lead in water seems inevitable in New York, we are finding a way to reduce the lead concentration in our drinking water as much as possible. Minnesota Department of Health suggested that we should let the water run for some time before using water for drinking and cooking. They stated that it is an efficient way to reduce our exposure to lead in water because the more time water has been staying in pipes, the higher the lead concentration it. 

We use the "Free Residential at-the-tap Lead and Copper Data" dataset to test this hypothesis. Each residential volunteer submitted two water samples: one is collected right after the tap is opened and the other is collected after the tap has flushed for 1-2 minutes. We use a ridgeline plot to compare the lead concentration between two types of water samples. The plot gives us two density curves with lead concentration of water samples on the x-axis and density on the y-axis, one for first draw samples and the other for 1-2 minute draw. We can observe that among 1-2 minutes draw samples,the percentage of water samples with lead concentration around 0 is higher than that of first draw samples. On the contrary, the percentage of water samples with large lead concentration is higher in 1-2 minutes draw samples. Therefore, we can conclude that lead concentration in water will decrease as the tap is opening and we can reduce our exposure to lead in drinking water by waiting for some time after opening the tap. 

```{r}
tap_rate <- read_csv("tap_rate_by_zipcode.csv") %>% mutate(Zipcode = as.character(Zipcode))
tap_rate_residential <- tap_rate %>% filter(source == "free residential") %>% arrange(rate) %>% top_n(10)
tap_rate_compliance <- tap_rate %>% filter(source == "compliance") %>% arrange(rate) %>% top_n(10)
tap_rate_bar <- bind_rows(tap_rate_residential, tap_rate_compliance)
ggplot(tap_rate_bar, aes(x = fct_inorder(Zipcode), y = rate)) +
  geom_bar(stat = "identity", fill = "#7C9B99") +
  facet_grid(rows = vars(source), scales = "free", space = "free") +
  ggtitle ("Areas with Relative High Rate of Unsafe Water Samples in NYC") +
  labs(x = "Zipcode", y = "Rate of unsafe samples with lead level over 0.015 mg/L") +
  coord_flip() + theme_grey()
```

To further investigate the prevalence of lead in tap water in NYC, we examined the difference of lead level in water samples over different areas across the city. Labeling the water samples with lead level over 0.015 mg/L as unsafe water, the above Bar Chart shows the proportion of unsafe water samples in areas represented by the unit of zip code. It summarizes the top 10 areas with high rate of unsafe water samples in NYC for both compliance and free residential datasets. The top rates of unsafe water samples in compliance data are greater than those in free residential data generally. There are no common areas with high rate in both datasets. It is probably because the compliance dataset has smaller sample size and these two data were collected in different methods for measurement. Among these areas, Greenwich Village has the highest rate of unsafe water samples, meaning that almost half of places with in this area are likely to be supplying unsafe tap water, which the government needs to further study the reasons behind.

```{r}
tap_avg <- read_csv("tap_avg_by_zipcode.csv") %>% mutate(Zipcode = as.character(Zipcode))
tap_avg_residential <- tap_avg %>% filter(source == "free residential") %>% arrange(avg_lead_level) %>% top_n(10)
tap_avg_compliance <- tap_avg %>% filter(source == "compliance") %>% arrange(avg_lead_level) %>% top_n(10)
tap_avg_bar <- bind_rows(tap_avg_residential, tap_avg_compliance)
ggplot(tap_avg_bar, aes(x = fct_inorder(Zipcode), y = avg_lead_level)) +
  geom_bar(stat = "identity", fill = "#00AFBB") +
  facet_grid(rows = vars(source), scales = "free", space = "free") +
  ggtitle ("Average Lead Level of At-the-tap Water Samples in NYC") +
  labs(x = "Zipcode", y = "Average lead level (mg/L)") +
  coord_flip() + theme_grey()
```
Then we focused on those water samples with lead level over 0.015 mg/L, defined as unsafe water samples as above. We took average over the lead level of all unsafe tap water samples in different areas for both datasets to investigate how unsafe the water sources in those areas are. In the above Bar Chart, there are still variance of top areas between both datasets but the areas with zip code 10304 and 11379 have extremely high level of lead in tap water, which is more than 10 times of the stated limit. The area with zip code 10304 also has high rate of unsafe water samples in the previous Bar Chart "Areas with Relative High Rate of Unsafe Water Samples in NYC". More attention should be pay to Stapleton area (zip code 10304) in NYC.

## How are children in NYC affected by lead poisoning?
```{r, warning=FALSE, message=FALSE}
children_bar <- children_6_yrs %>% rename(`5 <= BLL <10 (mcg/dL)` = Children.under.6.years.with.elevated.blood.lead.levels..BLL..Rate..BLL..5.µg.dL.per.1.000.tested, `10 <= BLL <15 (mcg/dL)` = Children.under.6.years.with.elevated.blood.lead.levels..BLL..Rate.BLL..10.µg.dL.per.1.000.tested, `BLL >=15 (mcg/dL)` = Children.under.6.years.with.elevated.blood.lead.levels..BLL..Rate.BLL..15.µg.dL.per.1.000.tested) %>%
  rename(year = time_period) %>%
  mutate(year = as.character(year)) %>%
  filter(geo_type =="Citywide") %>%
  select(year, `5 <= BLL <10 (mcg/dL)`, `10 <= BLL <15 (mcg/dL)`, `BLL >=15 (mcg/dL)`) %>%
  na.omit %>%
  group_by(year) %>% 
  summarise(`5 <= BLL <10 (mcg/dL)` = sum(`5 <= BLL <10 (mcg/dL)`), `10 <= BLL <15 (mcg/dL)` = sum(`10 <= BLL <15 (mcg/dL)`), `BLL >=15 (mcg/dL)` = sum(`BLL >=15 (mcg/dL)`)) %>% 
  mutate(`5 <= BLL <10 (mcg/dL)` = `5 <= BLL <10 (mcg/dL)` - `10 <= BLL <15 (mcg/dL)`, `10 <= BLL <15 (mcg/dL)` = `10 <= BLL <15 (mcg/dL)` - `BLL >=15 (mcg/dL)`) %>%
  pivot_longer(cols = !year, names_to = "range", values_to = "Number of children with BLL >= 5 per 1000 tested")

ggplot(data = children_bar, aes(x = year, y=`Number of children with BLL >= 5 per 1000 tested`, fill = factor(range, levels=c("BLL >=15 (mcg/dL)", "10 <= BLL <15 (mcg/dL)", "5 <= BLL <10 (mcg/dL)")))) +
  geom_bar(stat='identity') +
  theme(legend.title = element_blank()) +
  scale_fill_manual(values = c("#F8766D","#7C9B99", "#00BFC4")) +
  ggtitle("Children’s Blood Lead Levels (BLL) in NYC") +
  theme(plot.title = element_text(lineheight=.8, face="bold", hjust = 0.5))
```

We subset our dataset to focus on the blood lead levels (BLL) of children living in New York City. The Centers for Disease Control and Prevention considers a blood lead level over 10 μg/dL abnormal but a level lower than 10 μg/dL might also be harmful. In the stacked bar chart above, we summarize in each year, the number of children with BBL >= 5 μg/dL per 1000 tested. We can observe that the number is decreasing every year. However, there are still many children harmed by lead every year. Therefore, we should attach great importance to the problem of lead contamination in water. 




```{r, warning=FALSE, message=FALSE}
children <- children_6_yrs %>%
  rename(`sample_size(in thousand)` = Children.under.6.years.with.elevated.blood.lead.levels..BLL..Number.Tested, `high_BLL_children_per_1000_tested(sum of 2005-2016)` = Children.under.6.years.with.elevated.blood.lead.levels..BLL..Rate..BLL..5.µg.dL.per.1.000.tested) %>%
  select(geo_area_name, `sample_size(in thousand)`, `high_BLL_children_per_1000_tested(sum of 2005-2016)`) %>%
  group_by(geo_area_name) %>%
  summarise(`sample_size(in thousand)` = sum(`sample_size(in thousand)`), `high_BLL_children_per_1000_tested(sum of 2005-2016)` = sum(`high_BLL_children_per_1000_tested(sum of 2005-2016)`))

children$`sample_size(in thousand)` = children$`sample_size(in thousand)` / 1000
children <- children %>% pivot_longer(col = !geo_area_name, names_to = "data_type", values_to = "value")
ggplot(children, aes(value, fct_reorder2(`geo_area_name`, `data_type`=="high_BLL_children_per_1000_tested(sum of 2005-2016)", value, .desc = FALSE), color = data_type)) +
  geom_point() + 
  ggtitle("Sample Size and High BBL Rate Grouped by Area") + 
  theme(legend.position="bottom") +
  theme(axis.text.y = element_text(size = rel(.75))) +
  ylab("") +
  theme(plot.title = element_text(lineheight=.8, face="bold", hjust = 0.5)) + 
  guides(color = guide_legend(nrow = 2))

```

From this Cleveland plots, we can figure out from 2005 to 2016, for each region in New York, how many children under 6 years old in total are tested lead concentration in blood, marked as blue, and among those tested, how many children in total has blood lead level over 5 μg/dL per 1000 tested in , marked as red. The graph will tell us how many children are affected by lead and whether the number of high BLL children per 1000 tested is related to the sample size. The sample size in different regions may differ because population sizes are different and test capabilities are different. We want to figure out if the difference in sample sizes affects the difference in ratio of children with high BLL.

In the graph, regions are sorted by the number of children with high BLL per 1000 tested in an decreasing order. We can observe that as the red points are ordered, blue points are scattered so there is no obvious pattern of relationship between sample size and number of high BLL children. Some regions such as New York City and Queens have large sample sizes but they can still have a low number of high BLL children per 1000 tested. Therefore, the numbers of high BLL children per 1000 tested from different regions are not biased by the difference in sample sizes.

From the plot, we can see that the high BLL rates in different regions differ. Some regions such as Greenpoint and Borough Park has high BLL rates while some regions such as Upper East Side and Willowbrook has much lower BLL rates than other regions. The reason might be that in some places the water pipes and solders contain more lead and government should notice that and replace some of them to guarantee the health of residents.

## Is there any relationship between the lead in NYC water system and the lead level in children's blood?

```{r}
relation <- read.csv("scatter_data.csv")
ggplot(relation, aes(rate_16, crate_16)) + geom_point(color = "#7C9B99") +
  labs(x = "Rate of unsafe water samples with lead level over 0.015 mg/L", y = "Number of high blood lead level samples per 1000 tests") +
  ggtitle("Potential Relationship between Lead in Water and in Children's Blood") +
  theme_light()
```

In order to explore the relationship betweeen the lead level in NYC water system and the lead level in children's blood, we join two datasets by areas. The scatterplot displays the relationship between the rate of unsafe water samples (lead level exceeding threshold) during 2014 to 2016 and the rate of high blood lead level samples in 2016. Each point in the plot represents a neighborhood. According to the scatter plot, there is no clear correlation between these two variables, which indicates that many other factors may influence the lead level in children's blood and the bad water quality is not the only cause. Further studies need to be done to investigate what result in the children's high lead level in blood.
