==> 06-interactive.Rmd <==
# Interactive component

<script src="https://d3js.org/d3.v6.js"></script>

<h2>Is there any relationship between at-the-tap water lead level and lead level in children's blood? </h2>

  <div id="my_dataviz">
    
  <p id="d2014">click here to see data of 2014<p>
  <p id="d2015">click here to see data of 2015<p>
  <p id="d2016">click here to see data of 2016<p>
  <p id="note">move your mouse over the dots to see name of the neighborhood!<p>

  </div>

<script src="scripts/scatter.js"></script>


The interactive scatter plot above shows the relationship between at-the-tap water lead level and blood lead level of children in NYC. Each dot represents a neighborhood, and each color represents a borough in NYC. You can explore the trend of data from 2014 to 2016 by clicking on the buttons above, you can also move your mouse over the dots in the graph to see the name of the neighborhood it represents.
